move getNextAction to util
