+++ lib/hooks/useGCFormContext.tsx
@@ -1,7 +1,13 @@
 "use client";
 import React, { createContext, useContext, ReactNode } from "react";
 import { PublicFormRecord } from "@lib/types";
-import { mapIdsToValues, FormValues, idArraysMatch, GroupsType } from "@lib/formContext";
+import {
+  mapIdsToValues,
+  FormValues,
+  idArraysMatch,
+  GroupsType,
+  getNextAction,
+} from "@lib/formContext";
 
 interface GCFormsContextValueType {
   updateValues: ({ formValues }: { formValues: FormValues }) => void;
@@ -37,36 +43,7 @@ export const GCFormsProvider = ({
 
     // @todo add a test for this
     if (hasNextAction(currentGroup)) {
-      let nextAction = groups[currentGroup].nextAction || "";
-
-      // Check to see if next action is an array
-      /*
-       i.e. multiple next actions based on choice
-       
-      "nextAction": [
-        { "choiceId": "1.0", "groupId": "f3e91cd0-343c-46a2-acab-3346865974cb" },
-        { "choiceId": "1.1", "groupId": "93278f8e-5d47-4507-a104-c47e8a950da0" }
-      ]
-      */
-      if (Array.isArray(nextAction)) {
-        let matched = false;
-
-        // Check for catch-all value
-        const catchAllRule = nextAction.find((action) => action.choiceId === "catch-all");
-
-        nextAction.forEach((action) => {
-          const match = matchedIds.includes(action.choiceId);
-          if (match) {
-            nextAction = action.groupId;
-            matched = true;
-          }
-        });
-
-        // If no match, try using catch-all
-        if (!matched && catchAllRule) {
-          nextAction = catchAllRule.groupId;
-        }
-      }
+      const nextAction = getNextAction(groups, currentGroup, matchedIds);
 
       if (typeof nextAction === "string") {
         setCurrentGroup(nextAction);
